TASK: Move ReleaseService to new namespace
