+++ Classes/Package.php
@@ -7,6 +7,7 @@ namespace t3n\JobQueue\RabbitMQ;
 use Flowpack\JobQueue\Common\Job\JobManager;
 use Neos\Flow\Core\Bootstrap;
 use Neos\Flow\Package\Package as BasePackage;
+use t3n\JobQueue\RabbitMQ\Service\ReleaseService;
 
 class Package extends BasePackage
 {
